Extract logic into UserContext
